refactor: inline WithAmqpChannel trait into AmqpPublisher
Trait had a single implementation. Inlining removes an unnecessary
abstraction and one file with no behavior change.
