+++ src/main/scala/org/galaxio/gatling/amqp/client/AmqpPublisher.scala
@@ -1,12 +1,30 @@
 package org.galaxio.gatling.amqp.client
 
+import com.rabbitmq.client.{AlreadyClosedException, Channel, ShutdownSignalException}
 import io.gatling.core.session.Session
 import org.galaxio.gatling.amqp.protocol.AmqpComponents
 import org.galaxio.gatling.amqp.request._
 
-class AmqpPublisher(destination: AmqpExchange, components: AmqpComponents) extends WithAmqpChannel {
+class AmqpPublisher(destination: AmqpExchange, components: AmqpComponents) {
 
-  private val awaitConfirm = components.protocol.publisherConfirms
+  private val awaitConfirm             = components.protocol.publisherConfirms
+  private val pool: AmqpConnectionPool = components.connectionPublishPool
+
+  private def withChannel[T](channelAction: Channel => T): T = {
+    val ch = pool.channel
+    try {
+      val result = channelAction(ch)
+      pool.returnChannel(ch)
+      result
+    } catch {
+      case e @ (_: AlreadyClosedException | _: ShutdownSignalException) =>
+        pool.invalidate(ch)
+        throw e
+      case e: Throwable                                                 =>
+        pool.returnChannel(ch)
+        throw e
+    }
+  }
 
   def publish(message: AmqpProtocolMessage, session: Session): Unit = {
 
@@ -38,6 +56,4 @@ class AmqpPublisher(destination: AmqpExchange, components: AmqpComponents) exten
         }
     }
   }
-
-  override protected val pool: AmqpConnectionPool = components.connectionPublishPool
 }
